move examples to separate module
