+++ project/BuildHelper.scala
@@ -29,7 +29,7 @@ object BuildHelper {
     "-Ywarn-value-discard",
     "-unchecked",
     "-deprecation",
-//    "-Xfatal-warnings"
+    "-Xfatal-warnings"
   )
 
   private val stdOpts213 = Seq(
